doc: comment all `runopts.Captured` functionality (and move it to a separate file)
