+++ runopts/runopts.go
@@ -156,42 +156,3 @@ func GenesisAlloc(alloc types.GenesisAlloc) Option {
 		return nil
 	})
 }
-
-type Captured[T any] struct {
-	Val T
-
-	apply Func
-}
-
-var _ Option = (*Captured[struct{}])(nil)
-
-func (c *Captured[T]) Apply(cfg *Configuration) error {
-	return c.apply(cfg)
-}
-
-func Capture[T any](fn func(*Configuration) T) *Captured[T] {
-	c := new(Captured[T])
-	c.apply = func(cfg *Configuration) error {
-		c.Val = fn(cfg)
-		return nil
-	}
-	return c
-}
-
-func CaptureConfig() *Captured[*Configuration] {
-	return Capture(func(c *Configuration) *Configuration {
-		return c
-	})
-}
-
-func CaptureBytecode() *Captured[[]byte] {
-	return Capture(func(c *Configuration) []byte {
-		return c.Contract.Bytecode()
-	})
-}
-
-func CaptureStateDB() *Captured[vm.StateDB] {
-	return Capture(func(c *Configuration) vm.StateDB {
-		return c.StateDB
-	})
-}
